Tapped into the query translation mechanism
Previous iterations of this code had a hacked version of the send()
function, that took an AST argument.

This version uses a custom query translator that simply turns the
AST into a string, JSON-formatted.

This is exactly what the lens_beacon_service needs and understands
when running queries against Beacon.

So the Beacon implementation of send could be simplified slightly, since
it no longer needs to do any serialization.