Set up simple docker-compose
The original docker-compose.yml was intended for a Lens installation
that talks with Beam in order to communicate with sites hosting
Bridgeheads.

The Beacon implementation is much more simplistic, it does not use
Beam in its current incarnation.

So I have simplified the docker-compose file.

I have also added a lens_beacon_service container to the docker-compose.
This is the backend componment needed for communication with Beacon
sites.

I have removed the dev version of docker-compose.