Added variant search.
The original variant search was renamed to HGVS, because it really uses
this standard for describing variants in a single line.

A new variant search was added, that allows users to specify chromosome,
start and end coordinates of the variant plus what actually changed
(right now, only transition SNPs are allowed). I used nested Category
objects because that was quick to implement, but ideally, we would
write a bespoke GUI component for this functionality.

I also made naming a bit more consistent, e.g. replacing "patient"
with "individual".